MDS-4625 Updated core mine tsf specs
