+++ services/core-web/src/components/mine/Tailings/MineReportTailingsInfo.js
@@ -1,6 +1,8 @@
 import React from "react";
 import MineTailingsInfoTabs from "./MineTailingsInfoTabs";
 
-const MineReportTailingsInfo = () => <MineTailingsInfoTabs enabledTabs={["reports", "map"]} />;
+export const MineReportTailingsInfo = () => (
+  <MineTailingsInfoTabs enabledTabs={["reports", "map"]} />
+);
 
 export default MineReportTailingsInfo;
