+++ services/core-web/src/components/mine/Tailings/MineTailingsInfo.js
@@ -1,6 +1,6 @@
 import React from "react";
 import MineTailingsInfoTabs from "./MineTailingsInfoTabs";
 
-const MineTailingsInfo = () => <MineTailingsInfoTabs enabledTabs={["tsf"]} />;
+export const MineTailingsInfo = () => <MineTailingsInfoTabs enabledTabs={["tsf"]} />;
 
 export default MineTailingsInfo;
